add loggin names

+++ modules/300-prometheus/images/grafana-dashboard-provisioner/hooks/dashboard_provisioner.sh
@@ -51,10 +51,12 @@ function __main__() {
     return 0
   fi
 
+  malformed_dashboards=/tmp/malformed_dashboards
   for i in $(context::jq -r '.snapshots.dashboard_resources | keys[]'); do
     dashboard=$(context::get snapshots.dashboard_resources.${i}.filterResult)
     title=$(jq -rc '.definition | fromjson | .title' <<< ${dashboard} | slugify)
     folder=$(jq -rc '.folder' <<< ${dashboard})
+    name=$(jq -rc '.name' <<< ${dashboard})
 
     file="${folder}/${title}.json"
 
@@ -65,8 +67,18 @@ function __main__() {
     fi
 
     mkdir -p "/tmp/dashboards/${folder}"
-    jq -rc '.definition' <<< ${dashboard} > "/tmp/dashboards/${file}"
+    jq -rc '.definition' <<< ${dashboard} > "/tmp/dashboards/${file}" 2> "err-${name}"
+    if [[ -f "err-${name}" ]]; then
+        rm "err-${name}"
+        echo "${name}" >> ${malformed_dashboards}
+    fi
   done
+
+  if [[ -f /tmp/malformed_dashboards ]]; then
+    names=$(echo -n "$(tr -s '\n' ', ' < "${malformed_dashboards}")" | sed 's/,$//')
+    >&2 echo "Malformed dashboard definitions: ${names}"
+    exit 1
+  fi
 
   clear_data
   cp -TR /tmp/dashboards/ /etc/grafana/dashboards/
